requirements.txt
face_recognition
Pillow
numpy
scikit-image
